requirements.txt
asgiref==3.3.1
azure-common==1.1.26
azure-core==1.10.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.5
django-ckeditor==6.0.0
django-countries==7.0
django-js-asset==1.2.2
django-storages==1.11.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
mysqlclient==2.0.3
oauthlib==3.1.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==8.1.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
whitenoise==5.1.0